pyproject.toml
[project]
name = "memoryagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.3",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.6.6",
    "yml>=0.0.1",
]
